Split up EmitsAuthenticationEvents
